backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
qdrant-client==1.15.1
fastembed==0.7.3
requests==2.32.5
httpx==0.28.1
PyYAML==6.0.2
beautifulsoup4==4.13.4
lxml==5.4.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
